requirements.txt
fastapiuvicornhttpxpython-dotenvopenaipillow